work with jwcdRequest
edit in ddl and dml for jwcdRequest and related tables
create from controller for repository
and and edit in some entities and mapper

not every thing are test good will complete some things

+++ backend/judicial-warrant/src/main/java/com/informatique/gov/judicialwarrant/support/modelmpper/CandidateMapper.java
@@ -19,7 +19,7 @@ import lombok.AllArgsConstructor;
 public class CandidateMapper extends AbstractModelMapper<Candidate, CandidateDto, Long> {
 
 	private ModelMapper<CandidateStatus, CandidateStatusDto, Byte> candidateStatusMapper;
-	private ModelMapper<Request, RequestDto, Integer> requestMapper;
+	private ModelMapper<Request, RequestDto, Long> requestMapper;
 	private ModelMapper<OrganizationUnit, OrganizationUnitDto, Short> organizationUnitMapper;
 	private ModelMapper<CandidateHistoryLog, CandidateHistoryLogDto, Long> candidateHistoryLogMapper;
